requirements.txt
# HomePilot - Edge AI Voice Assistant Dependencies
# ================================================

# Wake Word Detection
pvporcupine>=3.0.0
pvrecorder>=1.2.0

# Speech-to-Text (Offline)
vosk>=0.3.45

# Text-to-Speech (Offline)
piper-tts>=1.2.0

# Audio Processing
sounddevice>=0.4.6
numpy>=1.24.0
soundfile>=0.12.0

# Home Assistant Integration
aiohttp>=3.9.0

# Security & Cryptography
cryptography>=41.0.0

# Configuration
pyyaml>=6.0.1

# Scheduling (Timers)
apscheduler>=3.10.0

# System Utilities
psutil>=5.9.0
